% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherScripts.R
\name{Add.NA}
\alias{Add.NA}
\title{Add NA Values Function}
\usage{
Add.NA(VEC, by = 60)
}
\arguments{
\item{VEC}{Xts vector.}

\item{by}{Time step of xts vector in seconds. Default is one minute (60 s).}
}
\value{
Xts vector with uniform time step.
}
\description{
The function unifies the time step by adding NA values to gaps.
}
\examples{
TS=DummyTS()
TS2=CutTimeSeriesDel(TS,start="2016-01-02 12:00:00",end="2016-01-02 22:00:00")
TS2=Add.NA(TS2)
DynPlot(cbind(TS,TS2))

}
