% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manning.R
\name{AEGG}
\alias{AEGG}
\title{An Area Defining Function (Egg Profile)}
\usage{
AEGG(w, r)
}
\arguments{
\item{w}{Decimal [m]. Water level of the egg profile.}

\item{r}{Decimal [m]. Radius of the egg profile.}
}
\value{
Returns the area of flow cross section [m^2].
}
\description{
This function analytically calculates the area of a flow cross section for a water level.\cr
Mind the units!!\cr
If you get errors using this function please make sure no NA values are within the water level vector and the waterlevels are plausible. (No negative water levels or water levels higher than the profile.)
}
\examples{
A=AEGG(w=1,r=1)

}
\seealso{
\link[ISI.Toolbox]{ManningEGG}
}
\keyword{Manning}
