% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{AllFilters}
\alias{AllFilters}
\title{A NA Deletion Function}
\usage{
AllFilters(XTS)
}
\arguments{
\item{XTS}{xts vector to be freed of NA's.}
}
\value{
The cleaned xts.
}
\description{
The function utilizes the DeleteNAs(), DeleteNATimes() and DeleteDuplicteTime() functions in that order
}
\examples{
TS=AllFilters(TS)

}
\seealso{
\link[ISI.Toolbox]{DeleteNATimes},\link[ISI.Toolbox]{DeleteDuplicteTime}
}
\keyword{delete}
