% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manning.R
\name{AREC}
\alias{AREC}
\title{An Area Defining Function (Rectangular Profile)}
\usage{
AREC(w, b)
}
\arguments{
\item{w}{Decimal or numeric vector [m]. Water level of the rectangular profile.}

\item{b}{Decimal [m]. Width of the rectangular profile.}
}
\value{
Returns the area of flow cross section [m^2].
}
\description{
This function analytically calculates the area of a flow cross section for a water level.\cr
Mind the units!!\cr
If you get errors using this function please make sure no NA values are within the water level vector and the waterlevels are plausible. (No negative water levels or water levels higher than the profile.)
}
\examples{
A=AREC(w=1,b=1)

}
\seealso{
\link[ISI.Toolbox]{ManningREC}
}
\keyword{Manning}
