% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cynthias_functions.R
\name{BaseFlow}
\alias{BaseFlow}
\title{Base Flow Index Function}
\usage{
BaseFlow(Flow_XTS)
}
\arguments{
\item{Flow_XTS}{Xts vector of flow hydrograph with minimum duration of five days. Ideally no NA values are present and the time step is even. (NA could be replaced with zeros if present. To unify the the time step the Add.Na function could be used.)}
}
\value{
A list.\cr
[[1]] -> Base Flow Index (BFI) as numeric\cr
[[2]] -> daily mean, five day minima, and baseflow as matrix\cr
}
\description{
Calculates Base Flow Index of a flow hydrograph (Gustard et al. 1992)
}
\examples{
TS=DummyTS(days=30)
BaseFlowTS=BaseFlow(TS)
}
\keyword{baseflow}
