% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherScripts.R
\name{Check_Fill_Quality}
\alias{Check_Fill_Quality}
\title{A function for checking the FillNaValues Output.}
\usage{
Check_Fill_Quality(Fillcount, Before_Fill, FillReturnList)
}
\arguments{
\item{Fillcount}{Integer. This value indexes all filled days.}

\item{Before_Fill}{Xts Vector. Original data, before using the fill function.}

\item{FillReturnList}{List (returned from the FillNaValues function).}
}
\value{
A dynamic plot and the fit quality expressed as average coefficient of determination.
}
\description{
This function illustrates the fill quality of the best_fit method of the FillNaValues. A day at the time.
}
\examples{
TS=DummyTS(days=150)
for(i in 1:20){
 start=as.integer(runif(1,1,length(TS)))
 len=as.integer(runif(1,1,1440))
 TS[start:(start+len)]=NA
}
TS_filled=FillNaValues(TS,use = "best_fit",Gaprange = c(0,80),cn_exclude_last_row = FALSE)
Check_Fill_Quality(1,TS,TS_filled)

}
\seealso{
\link[ISI.Toolbox]{FillNaValues}
}
