% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{ConL_Exp_Pot}
\alias{ConL_Exp_Pot}
\title{Coefficient Extraction from Regressionobject.}
\usage{
ConL_Exp_Pot(regressionobject)
}
\arguments{
\item{regressionobject}{Regressionobject from POLRegression or LOGRegression.}
}
\value{
Returns confidence intervals coefficients from regressionobject with transformation.
}
\description{
Extracts the confidence intervals of coefficients of a regressionobject from POTRegression or EXPRegression
}
\examples{
x=seq(1,10,1)
y=exp(x)+seq(10,100,10)
POT=POTRegression(x,y)
print(ConL_Exp_Pot(POT))
}
\seealso{
\link[stats]{lm}, \link[ISI.Toolbox]{POTRegression},\link[ISI.Toolbox]{EXPRegression}
}
\keyword{Regression}
\keyword{lm,}
