% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{ConL_Poly_Log}
\alias{ConL_Poly_Log}
\title{Extraction Function.}
\usage{
ConL_Poly_Log(regressionobject)
}
\arguments{
\item{regressionobject}{Regressionobject from POLRegression or LOGRegression.}
}
\value{
Returns confidence intervals of coefficients from regressionobject without any transformation.
}
\description{
Extracts the confidence intervals of coefficients of a regressionobject from POLRegression or LOGRegression.
}
\examples{
x=seq(1,10,1)
y=exp(x)+seq(10,100,10)
POL=POLRegression(x,y)
print(ConL_Poly_Log(POL))
}
\seealso{
\link[stats]{lm}, \link[ISI.Toolbox]{POLRegression},\link[ISI.Toolbox]{LOGRegression}
}
\keyword{Regression}
\keyword{lm,}
