% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{CorrectOffset}
\alias{CorrectOffset}
\title{A function for correcting offsets.}
\usage{
CorrectOffset(xts, start, end, offset)
}
\arguments{
\item{TS}{Xts. A time series to be corrected.}
}
\value{
TS Xts. A time series with the corrections.
}
\description{
A certain time period can be selected and corrected.
}
\examples{
TS=DummyTS(days=30,Rain_off = TRUE)
TS=rbind(CutTimeSeries(TS,start(TS)+24*3600,start(TS)+48*3600)+30,CutTimeSeriesDel(TS,start(TS)+24*3600,start(TS)+200*3600))
TS_corr=CorrectOffset(TS,start(TS)+24*3600,start(TS)+200*3600,-30)
DynPlot(cbind(TS,TS_corr))

}
