% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tobias_functions.R
\name{CS}
\alias{CS}
\title{Dissolved Oxygen Saturation Function}
\usage{
CS(Temp, Pressure)
}
\arguments{
\item{Temp}{Water-Temperature in °C}

\item{Pressure}{Water-Surface-Level Atmospheric Pressure in kPa}
}
\value{
A numeric vector\cr
}
\description{
This function uses the APHA-Method to calculate maximum dissolved oxygen concentration in a open freshwater, from given water-temperature and water-surface-level atmospheric pressure.\cr
For more details see: American Public Health Association Eaton, A.D., 2006. Standard methods for the examination of water & wastewater, 21st ed. 2005, centennial ed., [Nachdr.]. ed. American Public Health Association, Washington, DC.\cr
}
\examples{
pressure=rnorm(n = 21,mean = 101.32,sd = 2)
temperature=sin(x=seq(4,10,by = 0.3))*5+15
cs=CS(temperature,pressure)

}
\keyword{APHA,}
\keyword{Oxygen}
\keyword{Saturation}
