% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{CutTimeSeries}
\alias{CutTimeSeries}
\title{Function for cutting an xts vector}
\usage{
CutTimeSeries(TimeSeries, start, end)
}
\arguments{
\item{TimeSeries}{xts vector to be cut.}

\item{start}{The new start of the xts (format: "\%Y-\%m-\%d \%H:\%M")}

\item{end}{The new end of the xts (format: "\%Y-\%m-\%d \%H:\%M")}
}
\value{
xts vector cut to the desired time window.
}
\description{
The function cuts an xts to a new time window
}
\examples{
XTS = CutTimeSeries(XTS, "2016-06-16 12:00", "2016-08-05 00:00")

}
\seealso{
\link[ISI.Toolbox]{CutTimeSeriesDel}
}
\keyword{cut}
