% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{CutTimeSeriesDel}
\alias{CutTimeSeriesDel}
\title{Function for cutting an xts vector}
\usage{
CutTimeSeriesDel(TimeSeries, start, end)
}
\arguments{
\item{TimeSeries}{xts vector to be cut.}

\item{start}{The start of the window to be deleted (format: "\%Y-\%m-\%d \%H:\%M")}

\item{end}{The end of the window to be deleted (format: "\%Y-\%m-\%d \%H:\%M")}
}
\value{
xts vector without the cut time window.
}
\description{
The function deletes a time window of an xts vector
}
\examples{
XTS = CutTimeSeriesDel(XTS, "2016-06-16 12:00", "2016-08-05 00:00")

}
\seealso{
\link[ISI.Toolbox]{CutTimeSeries}
}
\keyword{cut,}
\keyword{delete}
