% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{Decompose4TXS}
\alias{Decompose4TXS}
\title{Function for decomposing an xts vector}
\usage{
Decompose4TXS(XTS, decomp = c("daily", "weekly", "monthly"),
  sampletime = c("1-min", "5-min", "15-min", "1-hour"),
  typ = c("additive", "multiplicative"))
}
\arguments{
\item{XTS}{xts vector to be decomposed.}

\item{decomp}{A string. "daily","weekly" or "monthly" are possible inputs. Default is daily.}

\item{sampletime}{A string. "1-min","5-min", "15-min" or "1-hour" are possible inputs. Default is 1-min.}

\item{typ}{A string. "additive" or "multiplicative" or valid inputs. Default is "additive".}
}
\value{
Decompose object.
}
\description{
The function uses the standard decompose tool of the stats package and applies it to an xts object.
For a more detailed description please see the decompose function.
}
\examples{
x=DummyTS(Csd=1.3)
A=Decompose4TXS(x)
plot(A)
}
\seealso{
\link[stats]{decompose}
}
\keyword{Filter,}
\keyword{decompose}
\keyword{smooth,}
