% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{DeleteDuplicteTime}
\alias{DeleteDuplicteTime}
\title{A Deletion Function}
\usage{
DeleteDuplicteTime(vector, Print = TRUE, return_index = FALSE)
}
\arguments{
\item{vector}{xts to be freed of duplicate values in timestamps.}

\item{Print}{Boolean. Print the number of duplicate timestamps deleted?}

\item{return_index}{Boolean. If TRUE returns a list. First position is the filtered time series. Second position is a vector containing the indexes of the filtered values.}
}
\value{
Vector or xts without duplicate values in timestamp or depending on the return_index paramter a list.
}
\description{
Deletes all duplicate timestamps that occur in an xts
}
\examples{
TimeSeries=DeleteDuplicteTime(TimeSeries)

}
\keyword{NA}
\keyword{delete,}
