% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{DeleteNAs}
\alias{DeleteNAs}
\title{A NA Deletion Function}
\usage{
DeleteNAs(vector, Print = TRUE, return_index = FALSE)
}
\arguments{
\item{vector}{numeric or xts to be freed of NA's.}

\item{Print}{Boolean. Print the number of NA's deleted?}

\item{return_index}{Boolean. If TRUE returns a list. First position is the filtered time series. Second position is a vector containing the indexes of the filtered values.}
}
\value{
Vector or xts without NA - values or depending on the return_index paramter a list.
}
\description{
Deletes all NA - values of an xts or numeric (is equivalent to na.omit but prints the amount of NA's deleted)
}
\examples{
TimeSeries=DeleteNAs(TimeSeries)

}
\seealso{
\link[stats]{na.omit}
}
\keyword{NA}
\keyword{delete,}
