% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{DeleteNATimes}
\alias{DeleteNATimes}
\title{A NA Deletion Function}
\usage{
DeleteNATimes(vector, Print = TRUE, return_index = FALSE)
}
\arguments{
\item{vector}{xts to be freed of NA - timestamps.}

\item{Print}{Boolean. Print the number of NA - timestamps deleted?}

\item{return_index}{Boolean. If TRUE returns a list. First position is the filtered time series. Second position is a vector containing the indexes of the filtered values.}
}
\value{
Vector or xts without NA - values in timestamp or depending on the return_index paramter a list.
}
\description{
Deletes all NA - values that occur in the timestamp of an xts
}
\examples{
TimeSeries=DeleteNATimes(TimeSeries)

}
\keyword{NA}
\keyword{delete,}
