% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{DelZeroNegValues}
\alias{DelZeroNegValues}
\title{A Deletion Function}
\usage{
DelZeroNegValues(vector, sign = "<0", extendGap = 0)
}
\arguments{
\item{vector}{xts or numeric vector.}

\item{sign}{determines the relation that will be used for deleting. Possible inputs are "<0" or "<=0".}

\item{extendGap}{setting this paramter to a value greater 0 will utilize the ExtendFalse() function, thereby extending the resulting gaps in the data.}
}
\value{
The cleaned xts vector.
}
\description{
Deletes all zero and/or negative values of an xts or numeric
}
\examples{
TS=DummyTS()-15
TS2=DelZeroNegValues(TS,"<=0",1)
DynPlot(cbind(TS,TS2))

}
\seealso{
\link[ISI.Toolbox]{DummyTS},\link[ISI.Toolbox]{DynPlot},\link[ISI.Toolbox]{ExtendFalse}
}
\keyword{delete}
