% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreationFunctions.R
\name{DummyTS}
\alias{DummyTS}
\title{A Generating Function}
\usage{
DummyTS(days = 7, SampleTime_min = 1, sd_days = 0.25, Cunif = 0.01,
  Csd = 0.03, FouP = c(8.95652, -2.42785, -0.58782, -0.97698, -2.6713,
  0.34104, -0.14011, 0.10532, 0.57853, -0.15287, -0.15902),
  AnualPrecipitation_mm = 600, Anual_Days_of_P = 100,
  Beta = list(0.3, 0.5, 0.5, c(0.5, 0.5), c(2, 2), c(2, 2)),
  Lin_Stor_Catchment_Area = 2800,
  Lin_Stor_Number_of_Catchments_slow_fast = c(3, 1),
  Lin_Stor_Average_Lambda_slow_fast = c(1000, 100), Rain_off = FALSE,
  timezone = "UTC")
}
\arguments{
\item{days}{The number of days to be generated}

\item{SampleTime_min}{The logging interval in minutes. (has been tested for 1,2 and 5 minutes)}

\item{sd_days}{The standard deviation from day to day (the higher the greater the difference from day to day)}

\item{Cunif}{The uniform distributed noise to be set on the signal.}

\item{Csd}{The normal distributed noise to be set on the signal.}

\item{FouP}{A vector containing 11 values for the Fourier Series starting with a_0. The next coefficients are the amplitudes of cosine and sine terms.(every even index number is a cosine amplitude...)}

\item{AnualPrecipitation_mm}{The anual rainfall height in mm.}

\item{Anual_Days_of_P}{The number of days in rains in a year.}

\item{Beta}{A list containing 4 to 6 values of vectors that describe the probabilities of rain, for the rain's disaggregation. For a more detailed description see the Raingenerator() function.}

\item{Lin_Stor_Catchment_Area}{The total area used by the linear storage model. It is spread according to the relation between the slow and fast catchments set in the Lin_Stor_Number_of_Catchments_slow_fast vector.}

\item{Lin_Stor_Number_of_Catchments_slow_fast}{Vector of length 2. The first position represents the number of slow catchments and the second the fast ones.}

\item{Lin_Stor_Average_Lambda_slow_fast}{Vector of length 2. Specifying the average retardation coefficient of the slow and fast linear storage models. (First slow then fast.)}

\item{Rain_off}{Logical value. If TRUE the Raingenerator and the linear storage models won't be utilized.}

\item{timezone}{Timezone of returned xts.}
}
\value{
Returns a generated xts vector.
}
\description{
Generates an xts with an desired Dynamik. The default setting will create a fairly realistic discharge pattern.
The function achieves this buy using a Fourier Series of order 5 (the period is set to 2 pi), the Raingenerator function and multiple linear storage models.
For future reference slow catchments refer to the ones that do have a long rainfall concentration time, thus fast catchments do have a short rainfall concentration time.
}
\examples{
TS=DummyTS(days=30)
DynPlot(TS)

}
\seealso{
\link[ISI.Toolbox]{Raingenerator}
}
\keyword{create,}
\keyword{generate,}
\keyword{synthetic}
