% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotter.R
\name{DynPlot}
\alias{DynPlot}
\title{A Simple Dynamic Plotter based on Dygraphs}
\usage{
DynPlot(X, Title = NULL, Ylab = NULL, Y2lab = NULL, Labels = seq(1,
  dim(X)[2], 1), Dotsize = 1, YRange = NULL, Y2Range = NULL,
  colset = 2, Axis = rep(1, dim(X)[2]), Events = NULL,
  Eventnames = NULL, legendwidth = 200)
}
\arguments{
\item{X}{The xts to plot. For multiple use X=cbind(first,second,third)}

\item{Title}{String. Title of graph.}

\item{Ylab}{String. Title of left y-axis.}

\item{Y2lab}{String. Title of right y-axis.}

\item{Labels}{Vector of Strings. Length must be equal to dim(X)[2].}

\item{Dotsize}{Decimal.}

\item{YRange}{Vector of length 2. Setting the limits for the left y-axis.}

\item{Y2Range}{Vector of length 2. Setting the limits for the right y-axis.}

\item{colset}{A color palette. Possible inputs are: 1, 2 or 3.}

\item{Axis}{A vector. Length must be equal to dim(X)[2]. The vector specifies which y axis the xts should be plotted on. 1 equals the left axis, 2 equals the right axis.}

\item{Events}{A vector of datetime strings (format="\%Y-\%m-\%d \%H-\%M-\%S"). For each entry a vertical line will be added to the plot.}

\item{Eventnames}{A vector of strings. Must have same length as Events vector. Lables each event.}

\item{legendwidth}{Integer. Width of the legend shown in the dynamic plot. Default is 200.}
}
\value{
Dynamic Plot.
}
\description{
Plots data with less code than in dygraph with limited options.
}
\examples{
a=DummyTS(days=14)
b=DummyTS(days=14)
c=DummyTS(days=14)
DynPlot(cbind(a,b,c),Axis=c(1,2,2),Labels=c("First","Second","Third"),Events = c("2016-01-02 01:00:00"),Eventnames = c("Eventname"))
}
\seealso{
\link[dygraphs]{dygraph}
}
\keyword{dygraph,}
\keyword{dynamic}
\keyword{plot}
\keyword{plot,}
