% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{EraseTrue}
\alias{EraseTrue}
\title{A Function for manupulation of logical vectors}
\usage{
EraseTrue(LogiVec, UnplausibleGapsize = 2)
}
\arguments{
\item{LogiVec}{logical vector.}

\item{UnplausibleGapsize}{value up to which the TRUE sections will be replaced with FALSE sections}
}
\value{
manipulated logical vector
}
\description{
Erases the TRUE values of a logical vector if the TRUE sections are <= than a given value
}
\examples{
a=c(TRUE,TRUE, FALSE,TRUE,TRUE,FALSE,TRUE,TRUE,TRUE,TRUE,FALSE,TRUE,FALSE,TRUE,FALSE,TRUE,TRUE)
b=EraseTrue(a,1)

}
\keyword{logical}
