% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cynthias_functions.R
\name{EventIndex}
\alias{EventIndex}
\title{Hysteresis Index and First Flush Index Function}
\usage{
EventIndex(Event_XTS, k, FFx)
}
\arguments{
\item{Event_XTS}{Xts matrix of flow on the first column and concentration of any contaminant species on the second column. Ideally no NA values are present and the time step is even. (NA could be replaced with zeros if present. To unify the the time step the Add.Na function could be used.)}

\item{k}{numeric of the first segment of hysteresis and the equal interval for calculation of Hysteresis Index. For instance, k=0.1 means hysteresis is divided into 10 equal parts and HI is calculated for 9 segments (at 0.1,0.2,0.3,...,0.9); k=0.01 means hysteresis is divided into 100 equal parts and HI is calculated for 99 segments (at 0.01,0.02,0.03,...,0.99)}

\item{FFx}{numeric of the first discharge proportion where proportion of generated load to be determined. For instance, FFx=0.25 for FF25 criteria, FFx=0.30 for FF30 criteria.}
}
\value{
A list.\cr
[[1]] -> Hysteresis Index (HI) as numeric\cr
[[2]] -> First-Flush Index (FFx) as numeric\cr
}
\description{
Calculates Hysteresis Index after Lawler (2006) and First Flush Index of a single storm event hydrograph and concentration record.\cr
Hysteresis Index of each segment (k) is calculated as by comparing concentration measured at the rising limb and at the falling limb of the hydrograph. For more detail see Lawler, D., Petts, G., Foster, I., & Harper, S. (2006). Turbidity dynamics during spring storm events in an urban headwater river system: The Upper Tame, West Midlands, UK. Science of the Total Environment 360, 109-126.\cr
First Flush Index is defined as the amount of load generated by the first x\% of flow , e.g. FF25 criteria used the first 25\% of flow and FF30 criteria used the first 30\% of flow. The value is obtained by plotting the normalised cumulative load against normalised cumulative flow. For more details see Obermann, M., Rosenwinkel, K.-H., & Tournoud, M.-G. (2009). Investigation of first flushes in a medium-sized mediterranean catchment. Journal of Hydrology, 373, 405-415. doi:doi:10.1016/j.jhydrol.2009.04.038
}
\examples{
x=seq(0,7.9,0.1)
base_flow=0.2*sin(x)
flow_peak=8*dnorm(x = x,mean = 4)
base_tracer=0.7
tracer_peak=6*dnorm(x = x,mean = 3)
timestamp=seq(from=as.POSIXct("2018-01-01 00:00"),by="hour",length.out = 80)
flow=xts(base_flow+flow_peak,order.by=timestamp)
conc=xts(base_tracer+tracer_peak,order.by = timestamp)
data=cbind(flow,conc)
DynPlot(data,Axis = c(1,2))
stormindex=EventIndex(data,0.1,0.25)
}
\keyword{first}
\keyword{flush}
\keyword{flush,}
\keyword{hysteresis}
\keyword{hysteresis,}
\keyword{index}
\keyword{index,}
