% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{EXPRegression}
\alias{EXPRegression}
\title{Linear Regression for Exponential Function}
\usage{
EXPRegression(x, y, summary = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector. Must be same length like x.}

\item{summary}{Boolean. If TRUE summary will be plotted.}
}
\value{
Returns lm object (regressionobject).
}
\description{
Fits a exponential function: y=a*exp(b*x).\cr
Note that in order to use linear regression a transformation was used. The returned intercept value needs to be transformed back. use exp() or the function Par_Exp_Pot.
}
\examples{
x=seq(1,10,1)
y=exp(x)+seq(10,100,10)
EXP=EXPRegression(x,y)
plot(x,y)
lines(x,exp(predict(EXP)),col=2)
}
\seealso{
\link[stats]{lm}, \link[ISI.Toolbox]{Fitall}, \link[ISI.Toolbox]{Par_Exp_Pot}
}
\keyword{Regression}
\keyword{lm,}
