% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{ExtendFalse}
\alias{ExtendFalse}
\title{A Function for manupulation of logical vectors}
\usage{
ExtendFalse(vec, amount = 1)
}
\arguments{
\item{vec}{logical vector.}

\item{amount}{amount of extention of FALSE section.}
}
\value{
manipulated logical vector
}
\description{
Extends the FALSE sections of a logical vector by a desired amount.
}
\examples{
a=c(TRUE,TRUE, FALSE,FALSE,TRUE,TRUE,TRUE,TRUE,FALSE,TRUE,TRUE,TRUE,TRUE)
b=ExtendFalse(a,1)
c=ExtendFalse(a,2)

}
\keyword{logical}
