% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{FFT_XTS}
\alias{FFT_XTS}
\title{Function for Fast-Fourier-Transformation of an xts vector}
\usage{
FFT_XTS(TS, dt = 1/24/60, Return_FreqMag_Sin_Cos = FALSE,
  Return_Freq_Mag_Phase = FALSE, Plot_Freq_Mag = TRUE, Ylim = NULL,
  Xlim = NULL)
}
\arguments{
\item{TS}{Xts vector to be transformed.}

\item{dt}{The times tep in days. (One Minute time step -> 1/24/60)}

\item{Return_FreqMag_Sin_Cos}{Boolean. If TRUE the frequency vector and both the cosine and sine magnitudes seperately will be returned.}

\item{Return_Freq_Mag_Phase}{Boolean. If TRUE the frequency vector, magnitude vector and phase vector will be returned.}

\item{Plot_Freq_Mag}{Boolean. Returns the Frequency Magnitude plot.}

\item{Ylim}{Vector of length 2. Set the plot's y axis limits.}

\item{Xlim}{Vector of length 2. Set the plot's x axis limits.}
}
\value{
Depending on the set parameters Return_FreqMag_Sin_Cos, Return_Freq_Mag_Phase and Plot_Freq_Mag a list will be returned and possible a plot will show.\cr\cr
If Return_FreqMag_Sin_Cos and Return_Freq_Mag_Phase are TRUE:\cr
The list will contain 5 entires.\cr
[[1]] -> frequency vector\cr
[[2]] -> cosine magnitude vector\cr
[[3]] -> sine magnitude vector\cr
[[4]] -> total magnitude vector\cr
[[5]] -> phase vector\cr\cr
If Return_FreqMag_Sin_Cos is TRUE and Return_Freq_Mag_Phase is FALSE:\cr
The list will contain 3 entires.\cr
[[1]] -> frequency vector\cr
[[2]] -> cosine magnitude vector\cr
[[3]] -> sine magnitude vector\cr\cr
If Return_FreqMag_Sin_Cos is FALSE and Return_Freq_Mag_Phase is TRUE:\cr
The list will contain 3 entires.\cr
[[1]] -> frequency vector\cr
[[2]] -> total magnitude vector\cr
[[3]] -> phase vector\cr\cr
If Return_FreqMag_Sin_Cos and Return_Freq_Mag_Phase are FALSE:\cr
An empty list will be returned.\cr\cr
If Plot_Freq_Mag is TRUE the frequency magnitude plot will be shown.\cr
}
\description{
The function uses fft to transform the xts vector to the magnitude frequency domain. Note that the first magnitude in of a n fft corresponds to the average offset of the entire signal. The plot does not contain this value, the returned vectors do.
}
\examples{
x=DummyTS()
FFT_XTS(x)
}
\seealso{
\link[ISI.Toolbox]{OneEuroFilter}
}
\keyword{Fourier,}
\keyword{fft}
