% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{FFT_XTS_Filter}
\alias{FFT_XTS_Filter}
\title{Function for Fast-Fourier-Transformation based filtering of an xts vector}
\usage{
FFT_XTS_Filter(TS, dt = 1/24/60, max.mag = NULL, min.mag = NULL,
  freq.range = NULL, Xlim = NULL, Ylim = NULL)
}
\arguments{
\item{TS}{Xts vector to be filtered}

\item{dt}{The time step in days. (One Minute time step -> 1/24/60)}

\item{max.mag}{A number or a vector of length m/2. Magnitudes higher than max.mag will be erased from the signal.}

\item{min.mag}{A number or a vector of length m/2. Magnitudes lower than min.mag will be erased from the signal.}

\item{freq.range}{A vector of length m=2*n (n is element of positive integers). Specifiy a frequency range, for which max.mag, min.mag or both should be applied. Multiple ranges can be set at once.}

\item{Xlim}{Vector of length 2. Set the plot's x axis limits.}

\item{Ylim}{Vector of length 2. Set the plot's y axis limits.}
}
\value{
Returns the filtered xts object.
}
\description{
The function uses fft to transform the xts vector to the magnitude frequency domain. Limits across frequency and mangitudes can be set to filter certain parts of the signals.
}
\examples{
TS1=DummyTS(days = 100)
#This command doesnt filter only plot. It will return the unfiltered xts object.
A=FFT_XTS_Filter(x,dt = 1/24/60,Xlim = c(0,5))
#Different filter examples. The purpose of these examples is only to show possible filter settings, without any deeper context.
A=FFT_XTS_Filter(x,dt = 1/24/60,Xlim = c(0,5),max.mag = 9)
A=FFT_XTS_Filter(x,dt = 1/24/60,Xlim = c(0,5),min.mag = 9)
A=FFT_XTS_Filter(TS1,dt = 1/24/60,Xlim = c(0,1),freq.range = c(0.3,0.4),max.mag = 7)
A=FFT_XTS_Filter(TS1,dt = 1/24/60,Xlim = c(0,1),freq.range = c(0.2,0.4),min.mag = 7)
A=FFT_XTS_Filter(TS1,dt = 1/24/60,Xlim = c(0,2),freq.range = c(0,0.4,0.9,1.1),min.mag = c(0,2),max.mag = c(4,99))

}
\keyword{Fourier,}
\keyword{fft,}
\keyword{filter}
