% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherScripts.R
\name{FillNaValues}
\alias{FillNaValues}
\title{A function for dealing with NA values in data.}
\usage{
FillNaValues(XTS, use = "linear", maxgap = 4, Regkey = "POT",
  Poldegree = 2, Fourierdegree = 5, Gaprange = c(0, 60),
  Allowed_NAs_in_Fill_Day = 0, Min_R = 0.85,
  cn_exclude_last_row = TRUE, Regression100Points = FALSE)
}
\arguments{
\item{XTS}{Xts vetor. Data set with NA values to be filled. Depending in the use parameter different dimensions are required (see description).}

\item{use}{String. Specifies the method for filling the NA values. \cr
Possible inputs are: "linear", "spline", "regression", "diurnal_pattern" or "best_fit"}

\item{maxgap}{Integer. Used by the linear and by the spline method. Specifies the maximum gap size that will be filled.}

\item{Regkey}{String. Used by the regression method. Specifies the function that will be fitted. \cr
Possible inputs are: "POT", "POL", "EXP" or "LOG"}

\item{Poldegree}{Integer. Specifies the degree of the polynomial function to be fitted. If Regkey=="POL" and use=="regression".}

\item{Fourierdegree}{Integer. Specifies the degree of the Fourier Series to be fitted. If use=="diurnal_pattern".}

\item{Gaprange}{Numeric vector of length two. Only used by the best_fit method. The vector specifies the size of gaps that will be filled. The 2 values represent gap size percentages of a day. \cr
E.g.: c(0,60) Gaps between the sizes of 0\% and 60\% of a day will be filled.}

\item{Allowed_NAs_in_Fill_Day}{Number. Only used by the best_fit method. Specifies the allowed percent of NA values in the fill day.}

\item{Min_R}{Number. Only used by the best_fit method. This values represents the minimal quality of the fit. If it is not reached a day won't be considered for filling a gap. \cr
If the dimension of XTS is greater than (m,2) the number will be compared to the average fit quality.}

\item{cn_exclude_last_row}{Boolean. Only used by the best_fit method. If TRUE the last row won't be used for comparison. (Must be TRUE if dim(XTS)==(m,1))}

\item{Regression100Points}{Boolean. Only used by the regression method. If TRUE the regression model will be calculated with 100 points only instead of all available data.\cr
The advantage is that the fit will be a better representation for areas with little measurement point density.}
}
\value{
If use != "best_fit" the na - filled - xts vector will be returned. \cr
Otherwise a list will be returned:
[[1]] -> na - filled - xts vector
[[2]] -> a numerical vector containing the information of the fill quality
[[3]] -> a list containing numericial vectors with the day to be filled and the fill day.
}
\description{
This function offers a variety of options to fill gaps in the data.
It incoorporates linear interpolation and spline interpolation which can be applied to xts data sets of any dimension.\cr
The regression option is designed for xts data sets of dimension (m,2), where the first column represents x in the second y.
A polynomial, potency, exponential or logarithmic function will be fitted to x and y and will be used to fill the gaps in y.\cr
The option diurnal_pattern requires an xts data set of dimension (m,1) for which a Fourier Series will be fitted und used to fill the gaps.\cr
best_fit is another method available. It compares days with gaps to days without gaps and uses the best fit, which is then equalized, to fill the gap. (The algorithms uses apply.daily)
Xts data sets may have the dimension (m,n), depending on the cn_exclude_last_row paramter n or (n-1) columns will be used for comparison (by correlation).
Either way the last columns gaps will be filled.
}
\examples{
TS=DummyTS(days=150)
for(i in 1:20){
 start=as.integer(runif(1,1,length(TS)))
 len=as.integer(runif(1,1,1440))
 TS[start:(start+len)]=NA
}
TS_filled=FillNaValues(TS,use = "best_fit",Gaprange = c(0,80),cn_exclude_last_row = FALSE)

}
\seealso{
\link[zoo]{na.approx}, \link[zoo]{na.spline}, \link[ISI.Toolbox]{Check_Fill_Quality}
}
