% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{Fitall}
\alias{Fitall}
\title{Linear Regression for Log,Pot,Exp,Pol Functions.}
\usage{
Fitall(X, Y, poly_order = 2)
}
\arguments{
\item{X}{Numeric vector.}

\item{Y}{Numeric vector. Must be same length like x.}

\item{poly_order}{Integer. Order of polynomial function.}
}
\value{
Returns a plot.
}
\description{
Fits all regression types except of fourier and plots the results. This function aims to give an overview to find the best fit.
}
\examples{
x=seq(1,10,1)
y=exp(x)+seq(10,100,10)
Fitall(x,y,poly_order=4)
}
\seealso{
\link[stats]{lm}, \link[ISI.Toolbox]{POLRegression},\link[ISI.Toolbox]{POTRegression},\link[ISI.Toolbox]{LOGRegression},\link[ISI.Toolbox]{EXPRegression}
}
\keyword{Regression}
\keyword{lm,}
