% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manning.R
\name{FitManning}
\alias{FitManning}
\title{A function for fitting Manning Strickler to a h-Q-realtion.}
\usage{
FitManning(Geometry = c("EGG", "CIR", "REC", "TRA"),
  kst_I_start = c(80, 0.001), kst_I_lwr = c(50, 1e-04),
  kst_I_upr = c(150, 0.01), modelfitmethod = "BFGS", h_vec, Q_vec,
  width, radius, alpha, beta)
}
\arguments{
\item{Geometry}{String that specifies the channels geometry. Valid options are 'EGG','CIR','REC' or 'TRA'.Default is "EGG".}

\item{kst_I_start}{Numerical vector of length 2. The fitting algorithm requires all parameters to be in a vector. First position kst value [m^{1/3}/s] and second position I [-] (Slope).}

\item{kst_I_lwr}{Numerical vector of length 2. Lower boudries for kst and I.}

\item{kst_I_upr}{Numerical vector of length 2. Upper boudries for kst and I.}

\item{modelfitmethod}{String specifying the fitting algorithm to use. Options can be found under \link[FME]{modFit} for the parameter method.}

\item{h_vec}{Numerical vector containing the water level [m].}

\item{Q_vec}{Numerical vector containing the flow [m^{3}/s].}

\item{width}{Number. Width of geometry. Only Relevant for "REC" and "TRA".}

\item{radius}{Number. Radius of geometry. Only Relevant for "EGG" and "CIR".}

\item{alpha}{Decimal in rad. Angle of slope -> right (mathematically negative). Alpha=0 results in a vertical line. Only relevant for "TRA".}

\item{beta}{Decimal in rad. Angle of slope -> left (mathematically positive). Beta=0 results in a vertical line. Only relevant for "TRA".}
}
\value{
Returns the modFit object from the FME package. Use $par to extract the fitted parameters. You can also use indexing, depending on the modelfitmethod the first or the second postions contains the parameters.
}
\description{
This tool uses the FME modFit function to fit Manning Strickler to h,Q data. There are different fitting methods available (\link[FME]{modFit}).\cr
Mind the units, when using this function!!\cr
In order to get a realistic fit for the kst and the I parameters of Manning please give plausible lower and upper boundries. \cr
If you get errors using this function please make sure no NA values are within the water level vector and the waterlevels are plausible. (No negative water levels or water levels higher than the profile.)
}
\examples{
# creating H,Q with set values for I=0.001 and kst=100
H=runif(100,0,1.6)
Q=ManningEGG(h=H,r=0.8,I=0.001,kst=100)

# fitting Manning to data
Mod=FitManning(Geometry = "EGG",kst_I_start = c(80,0.0008), kst_I_lwr = c(70,0.0005), kst_I_upr = c(110,0.003),h_vec = H,Q_vec = Q,r=0.8)

# extraction the kst and the I value:
kstI=Mod$par

# sorting the waterlevel values in order to get a nice line plot
Hfit=sort(H)
# calculating Q with the fitted parameters
Qfit=ManningEGG(h=Hfit,r=0.8,I=kstI[2],kst=kstI[1])

# plotting original data (black dots) and the fitted data (red line)
plot(H,Q)
lines(Hfit,Qfit,col=2)

# As shown in this example, there are multiple solutions for the kst and I paramters.
# Although the values fitted aren't the same as the ones given for this example, the function lies within the data points perfectly.
}
\seealso{
\link[FME]{modFit},\link[ISI.Toolbox]{ManningMaul},\link[ISI.Toolbox]{ManningEGG},\link[ISI.Toolbox]{ManningCIR},\link[ISI.Toolbox]{ManningREC},\link[ISI.Toolbox]{ManningTRA}
}
\keyword{Manning,}
\keyword{Regression,}
\keyword{h-Q}
