% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{FixDrift}
\alias{FixDrift}
\title{Function for fixing linear drift in xts data}
\usage{
FixDrift(XTS_to_fix, Fix_Info)
}
\arguments{
\item{XTS_to_fix}{xts vector to be corrected from drift.}

\item{Fix_Info}{an xts vector that specifies the drift times and values. It consists of n+1 lines for n being the number of drift sections.
Each line contains a datetime value followed by the drift value and the value it will be corrected to. The first line contains the begin of the first drift and contains the two values 0,0 as there is no drift present at this time.
the next line marks the end of the first drift and the two values for correcting it. Further lines use the previous datetime as start for its drift and the current one as end of the drift.}
}
\value{
Xts vector without drift.
}
\description{
This function calculates linear trends according to the given data. These trends are used to deduct the drift.
}
\examples{
x=DummyTS(Csd=1.3)
A=Decompose4TXS(x)
plot(A)
t=seq(as.POSIXct("2016-01-01 00:00:00",tz="UTC"),as.POSIXct("2016-01-02 00:00:00",tz="UTC"),by="min")
v=0.5*sin(seq(0,30*pi,length.out = length(t)))+1
drift=seq(0,2,length.out = 720)
v[361:1080]=v[361:1080]+drift
TS=xts(v,order.by = t)
#Fixinfo is probably more efficient to be read out of a txt file.
Fixinfo=rbind(cbind(xts(0,order.by = as.POSIXct("2016-01-01 06:00:00")),xts(0,order.by = as.POSIXct("2016-01-01 06:00:00"))),
              cbind(xts(3.5,order.by = as.POSIXct("2016-01-01 17:59:00")),xts(1.5,order.by = as.POSIXct("2016-01-01 17:59:00"))))
print(Fixinfo)

DynPlot(cbind(TS,FixDrift(TS,Fixinfo)),Labels = c("TS with Drift","TS without Drift"))
}
