% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherScripts.R
\name{FixPeriodicity}
\alias{FixPeriodicity}
\title{Repair Function}
\usage{
FixPeriodicity(TS)
}
\arguments{
\item{vec}{Xts vector to be ´fixed.}
}
\value{
Xts vector with uniform time step.
}
\description{
Fixes periodicity necessary for plotting with dygraphs by deleting NA values in time stamp and add na values for an even time step.
}
\examples{
TimeSeries=FixPeriodicity(TimeSeries)

}
\keyword{fix,}
\keyword{periodicity}
