% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{FourierPattern}
\alias{FourierPattern}
\title{Fourier Pattern from Regressionobject.}
\usage{
FourierPattern(timeseries, Regressionobject)
}
\arguments{
\item{timeseries}{xts vector.}

\item{Regressionobject}{Regressionobject from FourierRegression.}

\item{summary}{Boolean. If TRUE summary will be plotted.}
}
\value{
Returns xts vector with Fourier Pattern.
}
\description{
Calculates an xts vector containing the Fourier Pattern using xts vector and Fourier regressionobject.
}
\examples{
D2=DummyTS(days=14,Rain_off=TRUE)
F=FourierRegression(D2)
DP=FourierPattern(D2,F)
DynPlot(cbind(D2,DP))
}
\seealso{
\link[stats]{lm}, \link[ISI.Toolbox]{FourierRegression}
}
\keyword{Fourier}
\keyword{Fourier,}
\keyword{Regression,}
\keyword{Series,}
\keyword{diurnal}
\keyword{lm,}
\keyword{pattern}
