% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{Fourierperday}
\alias{Fourierperday}
\title{Function for calculating an average Fourier Sieres for each day}
\usage{
Fourierperday(TS, order = 4, accepltabledev = 25,
  MinDataPointsDaily = 400)
}
\arguments{
\item{TS}{Xts vector to be analysed.}

\item{order}{The order of the Fourier Series.}

\item{accepltabledev}{To be set in percent. When this function is used, it initially calculates a Fourier Series for the entire xts vector (resembling the average daily pattern). To ensure that the daily fitted patterns are close to the average pattern this value compares
the daily fitted series to the average. If the deviation is greater than the set one the calculated pattern won't be used.}

\item{MinDataPointsDaily}{To be able to estimate a representative daily pattern enough data points must be provided for each day. If the set value (minimum observations per day) isn't met the day will be left out.}
}
\value{
The function returns a list. Its first position contains the daily fitted Fourier Patterns while the second entry contains the diurnal residuals. The third position contains the regression object of the average Fourier Series.
}
\description{
The function uses a linear model to fit a Fourier Series for each day of an xts vector.
}
\examples{
Y=DummyTS(days=20,Rain_off=TRUE)
Y=Outliers(Y,5,4,keyword="long",maxlenoutlier=30)
D=Fourierperday(Y,4)
DD=D[[1]];WD=D[[2]]
DynPlot(cbind(Y,DD,WD),Labels=c("Data","Fourier","Residuals"))
}
\seealso{
\link[ISI.Toolbox]{FourierRegression}
}
\keyword{Filter,}
\keyword{Fourier}
\keyword{smooth,}
