% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{FourierRegression}
\alias{FourierRegression}
\title{Linear Regression for Fourier Series.}
\usage{
FourierRegression(timeseries, orderoftransformation = 5, summary = TRUE)
}
\arguments{
\item{timeseries}{xts vector.}

\item{orderoftransformation}{Integer. Order of Fourier Series (n).}

\item{summary}{Boolean. If TRUE summary will be plotted.}
}
\value{
Returns lm object (regressionobject).
}
\description{
Fits Fourier Series to xts vector. As the period is set to one day, this function is mainly to calculate diurnal patterns.
Fourier Series: y=a_0 +a_1*cos(p*1*t)+b_1*sin(p*1*t) +a_2*cos(p*2*t)+b_2*sin(p*2*t) +... +a_n*cos(p*n*t)+b_n*sin(p*n*t)
(orderoftransformation=n, periode = p --> is fix, set to 1 day)
}
\examples{
D2=DummyTS(days=14,Rain_off=TRUE)
F=FourierRegression(D2)
}
\seealso{
\link[stats]{lm}, \link[ISI.Toolbox]{FourierPattern}
}
\keyword{Fourier}
\keyword{Fourier,}
\keyword{Regression,}
\keyword{Series,}
\keyword{diurnal}
\keyword{lm,}
\keyword{pattern}
