% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherScripts.R
\name{Histo}
\alias{Histo}
\title{Event Duration Count Function}
\usage{
Histo(TS, limit, maxgapsize = 24)
}
\arguments{
\item{TS}{Xts vector to be analysed.}

\item{limit}{Decimal. Value to be exceeded.}

\item{maxgapsize}{Maximal tolerable size of temporal gap in data (in hours). If the gap is larger the event will considered to have ended.}
}
\value{
Returns numerical vector containing duration (in seconds) of exceedance events.
}
\description{
Calculates the duration and quantity of events exceeding the limit-value.
}
\examples{
TS=DummyTS()
vec=Histo(TS,limit=30)

}
