% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{hQ}
\alias{hQ}
\title{Waterlevel Discharge Relations via Regression.}
\usage{
hQ(TS, regressionobject, Type = "POL")
}
\arguments{
\item{TS}{xts vector containing waterlevels.}

\item{regressionobject}{A Regressionobject.}

\item{Type}{String. Type of the Regressionobject. Possible inputs are "POL", "POT", "EXP" or "LOG".}
}
\value{
Returns xts vector containing discharge values.
}
\description{
Function calculates a discharge time series using a water level time series and a previously calculated regressionobject (POT=POTRegression(x,y) -> POT is a regressionobject).(mind the units!)
}
\examples{
x=seq(1,10,1)
y=exp(x)+seq(10,100,10)
TS=xts(x,order.by=seq(strptime("19.08.2016 10:00","\%d.\%m.\%Y \%H:\%M",tz=""), strptime("19.08.2016 10:09","\%d.\%m.\%Y \%H:\%M",tz=""), "mins"))
P=POTRegression(x,y,summary=FALSE)
Q=hQ(TS,P,Type='POT')
}
\seealso{
\link[ISI.Toolbox]{hQConf}
}
\keyword{Regression}
\keyword{lm,}
