% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{hQConf}
\alias{hQConf}
\title{Confidence Bands for Waterlevel Discharge Relations via Regression.}
\usage{
hQConf(TS, regressionobject, Type = "POL")
}
\arguments{
\item{TS}{xts vector containing waterlevels.}

\item{regressionobject}{A Regressionobject.}

\item{Type}{String. Type of the Regressionobject. Possible inputs are "POL", "POT", "EXP" or "LOG".}
}
\value{
Returns list of length 2. \cr
[[1]] -> xts vector containing discharge values for lower confidence band.\cr
[[2]] -> xts vector containing discharge values for upper confidence band.
}
\description{
Function calculates 2 discharge time series (lower and upper confidence bands) using a waterlevel time series and Regressionobject.(mind the units!)
}
\examples{
x=seq(1,10,1)
y=exp(x)+seq(10,100,10)
TS=xts(x,order.by=seq(strptime("19.08.2016 10:00","\%d.\%m.\%Y \%H:\%M",tz=""), strptime("19.08.2016 10:09","\%d.\%m.\%Y \%H:\%M",tz=""), "mins"))
P=POTRegression(x,y,summary=FALSE)
hQC=hQConf(TS,P,Type='POT')
}
\seealso{
\link[ISI.Toolbox]{hQ}
}
\keyword{Regression}
\keyword{lm,}
