% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherScripts.R
\name{KDH}
\alias{KDH}
\title{Concentration Duration Occurance Function}
\usage{
KDH(TS, limits, timeclasses, maxgapsize = 24, old = FALSE)
}
\arguments{
\item{TS}{Xts vector to be analysed.}

\item{limits}{Numeric vector. Values to be exceeded.}

\item{timeclasses}{Numeric vector. Values to be set in days.}

\item{maxgapsize}{Maximal tolerable size of temporal gap in data (in hours). If the gap is larger the event will considered to have ended.}

\item{old}{Boolean. If TRUE the occurance as an event will be returned. If FALSE the occurance time based on the respective time class will be returned.}
}
\value{
Returns matrix (top line time in seconds, first row intensity, other occurance).
}
\description{
Calculates the occurance of exceedance for given concentration limit values and temporal (duration) classes.
}
\examples{
TS=DummyTS()
M=KDH(TS,limits=c(10,20,30,40,50,60),timeclasses=c(0.5/48,1/48.,1/24.,1/12.,1/8.,1/6.,1/3.,1))
print(M)
}
