% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{LOGRegression}
\alias{LOGRegression}
\title{Linear Regression for Logarithmic Function}
\usage{
LOGRegression(x, y, summary = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector. Must be same length like x.}

\item{summary}{Boolean. If TRUE summary will be plotted.}
}
\value{
Returns lm object (regressionobject).
}
\description{
Fits a logarithmic function: y=a+b*ln(x).
}
\examples{
x=seq(1,10,1)
y=exp(x)+seq(10,100,10)
LOG=LOGRegression(x,y)
plot(x,y)
lines(x,predict(LOG),col=2)
}
\seealso{
\link[stats]{lm}, \link[ISI.Toolbox]{Fitall}
}
\keyword{Regression}
\keyword{lm,}
