% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manning.R
\name{LuCIR}
\alias{LuCIR}
\title{A Wetted Perimeter Defining Function (Circular Profile)}
\usage{
LuCIR(w, r)
}
\arguments{
\item{w}{Decimal [m]. Water level of the circular profile.}

\item{r}{Decimal [m]. Radius of the circular profile.}
}
\value{
Returns the wetted perimeter of flow cross section [m].
}
\description{
This function analytically calculates the wetted perimeter of a flow cross section for a water level.\cr
Mind the units!!\cr
If you get errors using this function please make sure no NA values are within the water level vector and the waterlevels are plausible. (No negative water levels or water levels higher than the profile.)
}
\examples{
Lu=LuEGG(w=1,r=1)

}
\seealso{
\link[ISI.Toolbox]{ManningCIR}
}
\keyword{Manning}
