% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manning.R
\name{LuTRA}
\alias{LuTRA}
\title{A Wetted Perimeter Defining Function (Trapezoid Profile)}
\usage{
LuTRA(w, b, alpha, beta)
}
\arguments{
\item{w}{Decimal or numeric vector [m].}

\item{b}{Decimal [m]. Width of bottom part of the trapezoid profile.}

\item{alpha}{Decimal in rad. Angle of slope -> right (mathematically negative). Alpha=0 results in a vertical line.}

\item{beta}{Decimal in rad. Angle of slope -> left (mathematically positive). Beta=0 results in a vertical line.}
}
\value{
Returns the wetted perimeter of flow cross section [m].
}
\description{
This function analytically calculates the wetted perimeter of a flow cross section for a water level.\cr
Mind the units!!\cr
If you get errors using this function please make sure no NA values are within the water level vector and the waterlevels are plausible. (No negative water levels or water levels higher than the profile.)
}
\examples{
Lu=LuTRA(w=1,b=1,alpha=pi/4,beta=pi/5)

}
\seealso{
\link[ISI.Toolbox]{ManningTRA}
}
\keyword{Manning}
