% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manning.R
\name{ManningEGG}
\alias{ManningEGG}
\title{A Flow Calculation Function (Egg Profile)}
\usage{
ManningEGG(h, r, kst, I)
}
\arguments{
\item{h}{numeric vector containing water levels [m].}

\item{r}{Decimal [m]. Radius of the egg profile.}

\item{kst}{Decimal [m^{1/3}/s]. Friction coefficient according to Manning Strickler. (Rauheitsbeiwert nach Strickler).}

\item{I}{Decimal [-]. Slope of flow cross section.}
}
\value{
Returns numerical vector containing discharge values [m^3/s].
}
\description{
This function analytically calculates the discharge for a vector of water levels. Using the AEGG and the LuEGG functions.\cr
Mind the units!!\cr
If you get errors using this function please make sure no NA values are within the water level vector and the waterlevels are plausible. (No negative water levels or water levels higher than the profile.)
}
\examples{
H=runif(100,0,2.1)
Q=ManningEGG(h=H,r=1,I=0.001,kst=100)

}
\seealso{
\link[ISI.Toolbox]{ManningMaul},\link[ISI.Toolbox]{ManningCIR},\link[ISI.Toolbox]{ManningREC},\link[ISI.Toolbox]{ManningTRA},\link[ISI.Toolbox]{ManningSchacht}
}
\keyword{Manning}
