% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manning.R
\name{ManningMaul}
\alias{ManningMaul}
\title{A Flow Calculation Function (Maulprofil)}
\usage{
ManningMaul(h, I, kst, r, increment = 0.001)
}
\arguments{
\item{h}{numeric vector containing water levels [m].}

\item{I}{Decimal [-]. Slope of flow cross section.}

\item{kst}{Decimal [m^{1/3}/s]. Friction coefficient according to Manning Strickler. (Rauheitsbeiwert nach Strickler).}

\item{r}{Decimal [m]. Radius of the Maulprofil.}

\item{increment}{Decimal [m]. This value defines the height of the trapezoids used for approximating the Maulprofil. The result's precision rises the smaller the increment value. 0.001 m is the default value.}
}
\value{
Returns numerical vector containing discharge values [m^3/s].
}
\description{
This function numerically calculates the discharge for a vector of water levels. Using the MaulProfil and the IterMaul functions.\cr
Mind the units!!\cr
If you get errors using this function please make sure no NA values are within the water level vector and the waterlevels are plausible. (No negative water levels or water levels higher than the profile.)
}
\examples{
H=runif(100,0,1.5)
Q=ManningMaul(h=H,r=1,I=0.001,kst=100)

}
\seealso{
\link[ISI.Toolbox]{ManningEGG},\link[ISI.Toolbox]{ManningCIR},\link[ISI.Toolbox]{ManningREC},\link[ISI.Toolbox]{ManningTRA},\link[ISI.Toolbox]{ManningSchacht}
}
\keyword{Manning}
