% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manning.R
\name{ManningSchacht}
\alias{ManningSchacht}
\title{A Flow Calculation Function (For a Manhole)}
\usage{
ManningSchacht(h, kst, I, r = 0.6, b = 1.47, hlimit = 0.21)
}
\arguments{
\item{h}{numeric vector containing water levels [m].}

\item{kst}{Decimal [m^{1/3}/s]. Friction coefficient according to Manning Strickler. (Rauheitsbeiwert nach Strickler).}

\item{I}{Decimal [-]. Slope of flow cross section.}

\item{r}{Decimal [m]. Radius of circular bottom part of manhole.}

\item{b}{Decimal [m]. Width of manhole.}

\item{hlimit}{Decimal [m]. Height of transition between circular shaped bottom and straight wall.}
}
\value{
Returns numerical vector containing discharge values [m^3/s].
}
\description{
This function analytically calculates the discharge for a vector of water levels. Using the A_Schacht and the Lu_Schacht functions.\cr
Mind the units!!\cr
As backwater effects are likely for this cross section (depending on the following) the results will be less accurate with rising water level.\cr
If you get errors using this function please make sure no NA values are within the water level vector and the waterlevels are plausible. (No negative water levels or water levels higher than the profile.)
}
\examples{
H=runif(100,0,1.6)
Q=ManningSchacht(h=H,b=2,r=0.8,hlimit=0.21,I=0.001,kst=100)

}
\seealso{
\link[ISI.Toolbox]{ManningMaul},\link[ISI.Toolbox]{ManningEGG},\link[ISI.Toolbox]{ManningCIR},\link[ISI.Toolbox]{ManningREC},\link[ISI.Toolbox]{ManningTRA}
}
\keyword{Manning}
