% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manning.R
\name{ManningTRA}
\alias{ManningTRA}
\title{A Flow Calculation Function (Trapezoid  Profile)}
\usage{
ManningTRA(h, b, alpha, beta, kst, I)
}
\arguments{
\item{h}{numeric vector containing water levels [m].}

\item{b}{Decimal [m]. Width of bottom part of the trapezoid profile.}

\item{alpha}{Decimal in rad. Angle of slope -> right (mathematically negative). Alpha=0 results in a vertical line.}

\item{beta}{Decimal in rad. Angle of slope -> left (mathematically positive). Beta=0 results in a vertical line.}

\item{kst}{Decimal [m^{1/3}/s]. Friction coefficient according to Manning Strickler. (Rauheitsbeiwert nach Strickler).}

\item{I}{Decimal [-]. Slope of flow cross section.}
}
\value{
Returns numerical vector containing discharge values [m^3/s].
}
\description{
This function analytically calculates the discharge for a vector of water levels. Using the ATRA and the LuTRA functions.\cr
Mind the units!!\cr
If you get errors using this function please make sure no NA values are within the water level vector and the waterlevels are plausible. (No negative water levels or water levels higher than the profile.)\cr
Also the angle should be between 0 and pi/2. Excluding pi/2!
}
\examples{
H=runif(100,0,1)
Q=ManningTRA(h=H,b=2,I=0.001,kst=100,alpha=pi/3,beta=pi/3)

}
\seealso{
\link[ISI.Toolbox]{ManningMaul},\link[ISI.Toolbox]{ManningEGG},\link[ISI.Toolbox]{ManningCIR},\link[ISI.Toolbox]{ManningREC},\link[ISI.Toolbox]{ManningSchacht}
}
\keyword{Manning}
