% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tobias_functions.R
\name{nighttime.slope}
\alias{nighttime.slope}
\title{Nighttime-Slope-Methode
This function calculates Ecosystem Respiration (ER) [mgO2/L*d] and Reaeration Coefficient (kO2) [1/d] from a timeseries of measured DO an calculated DO-Saturation on a daily basis by Nighttime-Slope Method.\cr
First step is calculation of sunset and sunrise. If your river is shaded true sunset and sunrise (solar irradiance close to zero) might differ from astronomical times. Use the setdel and risedel parameter to adjust delay (in seconds).\cr
It is recommended to determine true sunset by modeling or measurement to improve accuracy of calculation.\cr
Next Step is to identify nighttime DO-Minima.\cr
Then temporal DO-Gradient (??DO/??t mgO2/L*d) and Saturationdeficit (DOsat-DO) in the time between sunset and DO-Minima is fitted by a linear function. Slope results kO2 and Intersept ER.\cr
The timeseries has to be continuous, evenly spaced and without NA's.\cr
Function uses NOAA sunrise sunset calculator.\cr
For more details see: Hornberger, G.M., Kelly, M.G., 1975. Atmospheric reaeration in a river using productivity analysis. Journal of the Environmental Engineering Division 101, 729-739.}
\usage{
nighttime.slope(data, setdel, risedel, Start, End, tstep, tz, lat, long,
  plot, plot_path)
}
\arguments{
\item{data}{xts-timeseries of DO (mg/L) DOsat (mg/L) and Temperature (°C) (data must be in given order)}

\item{setdel}{timedelay of sunset (seconds)}

\item{risedel}{timedelay of sunrise (seconds)}

\item{Start}{start time as "YYYY/MM/DD"}

\item{End}{end time as "YYYY/MM/DD"}

\item{tstep}{timestep of the timeseries (minutes)}

\item{tz}{timezone (sunrise sunset calculation) as "UTC"}

\item{lat}{latitude in decimal degrees}

\item{long}{longitude in decimal degrees}

\item{plot}{boolean to plot resulting linear function (creates a plot for every day)}

\item{plot_path}{path to save the plots, name is generated from date, as "C:/Users/Test/"}
}
\value{
A dataframe containing INT (ER), Slp (kO2), Rsquared of the linear function, Date, Mean Temperature for timeframe, and number of datapoints used for calculation\cr
}
\description{
Nighttime-Slope-Methode
This function calculates Ecosystem Respiration (ER) [mgO2/L*d] and Reaeration Coefficient (kO2) [1/d] from a timeseries of measured DO an calculated DO-Saturation on a daily basis by Nighttime-Slope Method.\cr
First step is calculation of sunset and sunrise. If your river is shaded true sunset and sunrise (solar irradiance close to zero) might differ from astronomical times. Use the setdel and risedel parameter to adjust delay (in seconds).\cr
It is recommended to determine true sunset by modeling or measurement to improve accuracy of calculation.\cr
Next Step is to identify nighttime DO-Minima.\cr
Then temporal DO-Gradient (??DO/??t mgO2/L*d) and Saturationdeficit (DOsat-DO) in the time between sunset and DO-Minima is fitted by a linear function. Slope results kO2 and Intersept ER.\cr
The timeseries has to be continuous, evenly spaced and without NA's.\cr
Function uses NOAA sunrise sunset calculator.\cr
For more details see: Hornberger, G.M., Kelly, M.G., 1975. Atmospheric reaeration in a river using productivity analysis. Journal of the Environmental Engineering Division 101, 729-739.
}
\examples{
First create a sample timeseries
datetime is our index
datetime=seq(from=as.POSIXct("2019-06-14 12:00"), by=60*5, to=as.POSIXct("2019-06-17 12:00"))
Dirunal DO, and Temperature is represented by a Sine wave
do=sin(x=seq(1,19,by = (18/(length(datetime)-1))))*3+10
temperature=sin(x=seq(1,19,by = (18/(length(datetime)-1))))*5+12
Atmospheric Pressure kPa is a normal distributed function with standard pressure (101.32 kPa) as mean and sd=1.
pressure=rnorm(length(datetime),101.32,1)
Calculate DO-Saturation from Temperature and Atmospheric Pressure
cs=CS(temperature,pressure)
Combine the index (datetime), DO, DOSat(cs) and Temperature to a xts Timeseries
ts=xts(x = cbind(do,cs,temperature),order.by = datetime)
Calculate ER and kO2 with nighttime slope method.
nighttime.slope(data = ts,Start = "2019/06/14",End = "2019/06/16",tstep = 5,tz = "UTC-1",lat=51.050407,long= 13.737262)
}
\keyword{APHA,}
\keyword{Oxygen}
\keyword{Saturation}
