% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreationFunctions.R
\name{Noise}
\alias{Noise}
\title{A Generating Function}
\usage{
Noise(TS, PercentageNoise, Noisespread, Overlap = FALSE, NoiseSections = 1)
}
\arguments{
\item{TS}{Xts vector in which the noise is to be inserted.}

\item{PercentageNoise}{The percent of the signal that should inherit noise.}

\item{Noisespread}{Standard Deviation of the noise.}

\item{Overlap}{Boolean that allows (TRUE) different noise sections to overlap or not(FALSE).
For PercentageNoise >50 Overlap should be set to TRUE, as the computation will otherwise take quite long.}

\item{NoiseSections}{How many noisy section should be created.}
}
\value{
Returns xts vector with noise.
}
\description{
The function inserts noise into a given xts. (dim(xts) = c(n,1))
}
\examples{
A = DummyTS(Rain_off = T)
B = Noise(A,15,1.5,NoiseSections=3)
DynPlot(cbind(B,A),Labels=c("TS_with_Noise","TS_without_Noise"))

}
\seealso{
\link[ISI.Toolbox]{DynPlot}
}
\keyword{create,}
\keyword{generate,}
\keyword{noise,}
\keyword{outlier}
