% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{NoiseFilter}
\alias{NoiseFilter}
\title{Function for removing noise of an xts vector.}
\usage{
NoiseFilter(TS, window_width, s_max = 99999999, lag = 1,
  Regression = FALSE, Polydegree = 1, Analyse = FALSE)
}
\arguments{
\item{TS}{xts vector to be cut.}

\item{window_width}{The width of the moving window.}

\item{s_max}{The maximal Standard Deviation that is allowed.}

\item{lag}{The lag from one moving window to the next.}

\item{Regression}{Boolean. If TRUE linear regression will be used for each moving window.}

\item{Polydegree}{The order of the polynomial function that will be fitted in each moving window.}

\item{Analyse}{Boolean. If TRUE a summary of the sd's of all moving windows will be plotted.}
}
\value{
Xts vector without noise.
}
\description{
The function filters noise using std within a moving window. The Regression parameter allows to fit polynoms in each moving window and the sd is calculated of the residuals.
This allows for reduced influence of trends that might be in the xts vector. Use the Analyse function to get an idea about the sd distribution of the time series and thereby setting the s_max values.
}
\examples{
#Creating xts with noise
D=DummyTS(days=7, SampleTime_min = 1, Rain_off = T)
D=Noise(D,15,1.5)

#Analysing
Dclean=NoiseFilter(D,window_width=20,Analyse=TRUE)

#Filtering
Dclean=NoiseFilter(D,s_max=0.6,window_width=20)

#Plotting
DynPlot(cbind(D,Dclean),Labels=c("TS_with_Noise","TS_without_Noise"))

}
\seealso{
\link[ISI.Toolbox]{GradFilter}
}
\keyword{Filter,}
\keyword{Noise}
