% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{OneEuroFilter}
\alias{OneEuroFilter}
\title{Function for smoothing an xts vector}
\source{
\url{http://cristal.univ-lille.fr/~casiez/1euro/}
}
\usage{
OneEuroFilter(x, freq = 120, mincutoff = 1, beta = 1, dcutoff = 1)
}
\arguments{
\item{x}{xts vector to be smoothed.}

\item{freq}{A measure on how many data point are in one cycle. For instance samples per day. (1440 is dt = 1 minute)}

\item{mincutoff}{The greater mincutoff the less lag and more sensitive (less memory).}

\item{beta}{The smaller beta the smoother.}

\item{dcutoff}{The greater mincutoff the less lag and more sensitive (less memory).}
}
\value{
xts vector without the cut time window.
}
\description{
The function is for smoothing an xts or numeric vector. It's based on the publication by Géry Casiez.
For detailed description on how to use the filter please read the paper.
}
\examples{
TS=DummyTS(Csd = 1.2)
TS2=OneEuroFilter(TS,freq=1440)
DynPlot(cbind(TS,TS2))
}
\seealso{
\link[ISI.Toolbox]{Smooth}
}
\keyword{cut,}
\keyword{delete}
