% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreationFunctions.R
\name{Outliers}
\alias{Outliers}
\title{A Generating Function}
\usage{
Outliers(TS, PercentageError, Outlierspread, keyword = "short",
  maxlenoutlier = 10, longspreadfactor = 0.04)
}
\arguments{
\item{TS}{Xts vector in which the outliers are to be inserted.}

\item{PercentageError}{The percent of the signal that should be outliers.}

\item{Outlierspread}{Standard Deviation of the Outliers}

\item{keyword}{String "short" or "long". If "short" simple Outliers are created. If "long" Outliers are set in a sequence so that a Offset is generated.}

\item{maxlenoutlier}{Restricts the max length of offsets.}

\item{longspreadfactor}{Defines the std within an offset (several Outliers in a row) as fraction of Outlierspread.}
}
\value{
Returns xts vector with outliers
}
\description{
The function inserts outlier into a given xts. (dim(xts) = c(n,1))
}
\examples{
A = DummyTS(Rain_off = T)
B = Outliers(A,5,4)
DynPlot(cbind(B,A),Labels=c("TS_with_Outliers","TS_without_Outliers"))

}
\seealso{
\link[ISI.Toolbox]{DynPlot}
}
\keyword{create,}
\keyword{generate,}
\keyword{noise,}
\keyword{outlier}
