% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{Par_Exp_Pot}
\alias{Par_Exp_Pot}
\title{Extraction Function.}
\usage{
Par_Exp_Pot(regressionobject)
}
\arguments{
\item{regressionobject}{Regressionobject from POTRegression or EXPRegression.}
}
\value{
Returns coefficients from regressionobject with transformation.
}
\description{
Extracts the coefficients of a regressionobject from POTRegression or EXPRegression.
}
\examples{
x=seq(1,10,1)
y=exp(x)+seq(10,100,10)
POT=POTRegression(x,y)
print(Par_Exp_Pot(POT))
}
\seealso{
\link[stats]{lm}, \link[ISI.Toolbox]{POTRegression},\link[ISI.Toolbox]{EXPRegression}
}
\keyword{Regression}
\keyword{lm,}
