% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotter.R
\name{PlotKDH2D}
\alias{PlotKDH2D}
\title{Plot Function}
\usage{
PlotKDH2D(M, savetofolder = NULL, ColBarRange = NULL,
  LogScale = FALSE)
}
\arguments{
\item{M}{Matrix from KDH function.}

\item{savetofolder}{String. To specify folder and name where plot should be saved.}

\item{ColBarRange}{Vector of length 2. Range for the levelplot colorbar.}

\item{LogScale}{Boolean. If TRUE the colorbar (and occurances) will be log scaled.}
}
\value{
Returns Levelplot.
}
\description{
The function plots the results of the KDH function in a levelplot.
}
\examples{
TS=DummyTS()
M=KDH(TS,limits=c(10,20,30,40,50,60),timeclasses=c(0.5/48,1/48.,1/24.,1/12.,1/8.,1/6.,1/3.,1))
PlotKDH2D(M)
}
\seealso{
\link[ISI.Toolbox]{KDH}
}
