% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{POLRegression}
\alias{POLRegression}
\title{Linear Regression for Polynomial Function}
\usage{
POLRegression(x, y, order = 2, summary = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector. Must be same length like x.}

\item{order}{Integer. The order of the polynomial function to be fitted. Must be smaller than length of x.}

\item{summary}{Boolean. If TRUE summary will be plotted.}
}
\value{
Returns lm object (regressionobject).
}
\description{
Fits a ploynomial function: y=a_0+a_1*x+a_2*x^2+...a_n*x^n (order=n)
}
\examples{
x=seq(1,10,1)
y=exp(x)+seq(10,100,10)
POL=POLRegression(x,y)
plot(x,y)
lines(x,predict(POL),col=2)
}
\seealso{
\link[stats]{lm}, \link[ISI.Toolbox]{Fitall}
}
\keyword{Regression}
\keyword{lm,}
