% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreationFunctions.R
\name{RainGen}
\alias{RainGen}
\title{A Generating Function}
\usage{
RainGen(Num_Days = 7, AnualP_mm = 600, Anual_Days_of_Rain = 121,
  sampletime_min = 1, beta = list(0.3, 0.5, 0.5, c(0.5, 0.5), c(2, 2), c(2,
  2)), Rain_per_day_xts = NULL)
}
\arguments{
\item{Num_Days}{The number of days to be generated.}

\item{AnualP_mm}{Anual precipitation in mm.}

\item{Anual_Days_of_Rain}{The number of days in rains in a year.}

\item{sampletime_min}{The logging interval of the resulting xts vector in minutes.}

\item{beta}{This list may have 4 to 6 positions. The values at each postion my be a number or a vector of 2.
The number represents the probability of each disaggreation slice to contain a rain event. By the same logic the vector contains
the 2 shape parameters of a beta distribution which will be sampled to determine the probility.

Example:
beta=list(0.33,0.5,c(2,2),c(1,5))
The disaggregation will be performed down to hourly time steps (24/3 -> 3*8/2 -> 6*4/2 -> 12* 2/2).
In the first disaggregation step the probability of each of the 3 8 hour steps to contain a rain event is 0.33.
In the second disarggregation step the probability of each of the 6 4 hour steps to contain a rain event is 0.5.
The following two disaggregation steps set probabilities by random sampling a beta distribution.}

\item{Rain_per_day_xts}{If NULL (default) th function will use the Num_Days, AnualP_mm and Anual_Days_of_Rain parameters to set up a vector containing daily rain events. Alternatively an xts with actual daily precipitation values may be set as input, which will be used for disaggregation instead.}
}
\value{
Returns a generated precipitation signal (xts vector).
}
\description{
The function disaggregates a diurnal pattern to a desired timestep. The disaggregation uses the beta parameter to determine the probabilities of rain for each disaggregation step. The first step breaks 24 hours into 3 times 8 hours. The following steps bisect each of the previous time intervals.
If the beta parameter contains 4 values the disaggregation will result in hourly values. 6 values will result in 15 min time steps.
Any desired sampletime below 15 min will be calculated by simply spreading it out out evenly.
The disaggragation matches a probability to each disaggregation step time interval. These will be multiplied to finally determine the rain probility (portion) of the final time steps.
}
\examples{
Rain_xts=RainGen()

}
\keyword{create,}
\keyword{generate,}
\keyword{rain}
