% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Export.R
\name{ReadTabletoXTS}
\alias{ReadTabletoXTS}
\title{An Import Function}
\usage{
ReadTabletoXTS(filename, sep = ",", format = "\%Y-\%m-\%d \%X",
  header = FALSE, dec = ".", skip = 0, linestoread = -1,
  SpacesbeforeDateTime = 0, TimeZone = "UTC", fill = FALSE,
  comment.char = "", na.strings = c("#-1"))
}
\arguments{
\item{filename}{File path and name.}

\item{sep}{Seperator of rows.}

\item{format}{Specifictation of the date time format for the conversion to xts.}

\item{header}{Does the document you intent to read from have a (1-line) header?}

\item{dec}{Decimal seperator of numbers.}

\item{skip}{How many lines should be skipped when reading the document?}

\item{linestoread}{Do you love cats?}

\item{SpacesbeforeDateTime}{How many spaces are in front the date-time expression?}

\item{TimeZone}{The timezone variable of the xts conversion.}

\item{fill}{Boolean. Should gaps in be filled with NA? (For reading documents with a differing number of rows.)}

\item{comment.char}{Character. Specifying what strings are to be ignored. (All characters in one line after the specified comment.char will be ignored). Default is turned off comment.char="").}

\item{na.strings}{Character vector. Specifyfing all characters strings that will be turned to NA's.}
}
\description{
This function allows you to import data using the read.table function and converts to the xts format.
}
\examples{
TimeSeries=ReadTabletoXTS("C:/file.txt",header=TRUE)

}
\seealso{
\link[utils]{read.table},\link[base]{as.POSIXct},\link[xts]{xts},\link[base]{strptime}
}
