% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{Separate.To.HyS.HyW}
\alias{Separate.To.HyS.HyW}
\title{A function for sepereating hydrological summer and winter.}
\usage{
Separate.To.HyS.HyW(TS)
}
\arguments{
\item{TS}{Xts. A time series to be seperated.}
}
\value{
A list. First position is hydrological summer xts. Second postion is hydrological winter xts.
}
\description{
The dates are divided by 01.05. and 01.11..
}
\examples{
TS=DummyTS(days=800,Rain_off = TRUE)
HH=Separate.To.HyS.HyW(TS)
HyS=HH[[1]]
HyW=HH[[2]]

}
