% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherScripts.R
\name{timetonumber}
\alias{timetonumber}
\title{Time Conversion Function}
\usage{
timetonumber(datetime)
}
\arguments{
\item{datetime}{Xts vector or POSIXct vector.}
}
\value{
Time in seconds from the 01.01.1970 01:00 as numeric vector.
}
\description{
Converts a time to number.
}
\examples{
TimeIndexinNumbers=timetonumber(TimeSeries)

}
