% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherScripts.R
\name{TimeZoneConverter}
\alias{TimeZoneConverter}
\title{A function changing Timezones.}
\usage{
TimeZoneConverter(XTS, TZorigin = "Europe/Berlin", TZdest = "UTC")
}
\arguments{
\item{XTS}{Xts vector. The timezone does not matter. It will be adjusted to the TZorigin.}

\item{TZorigin}{TimeZone string. Tzone(XTS) will be adjusted to the TZorigin.}

\item{TZdest}{TimeZone string. TimeZone that the xts will be converted to.}
}
\value{
Xts with the desired timezone.
}
\description{
This function enables you to change timezones. Thereby getting rid of winter and summer daylight savings.
}
\examples{
TS=DummyTS(days=365,Rain_off = TRUE,timezone = "Europe/Berlin")
print(tzone(TS))
TS=TimeZoneConverter(TS)
print(tzone(TS))
}
