% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Export.R
\name{WriteXTStoTXT}
\alias{WriteXTStoTXT}
\title{An Import Function}
\usage{
WriteXTStoTXT(TS, DirName = "C:/Work/R/output.txt")
}
\arguments{
\item{TS}{Xts to be written to a txt.}

\item{DirName}{Destination and name of the file to be safed.}
}
\description{
This function allows to export an xts object to a file. It's based on the write.zoo function. Please use that one if you want more input options.
}
\examples{
WriteXTStoTXT(TS,"C:/Work/R/output.txt")

}
\seealso{
\link[zoo]{write.zoo}
}
